requirements.txt
bs4==0.0.2
click==8.1.8
jinja2==3.1.5
Markdown==3.7
PyYAML==6.0.2
watchdog==6.0.0
